refactor: refactor service codes.
