+++ .env
@@ -1,5 +1,5 @@
-UPPERCASE_LETTERS=false
-LOWCASE_LETTERS=true
+UPPERCASE_LETTERS=true
+LOWERCASE_LETTERS=true
 NUMBERS=true
 SPECIAL_CHARACTERS=true
-PASSWORD_LENGTH=8
+PASSWORD_LENGTH=12
