+++ src/services/password/create.js
@@ -2,7 +2,7 @@ import chalk from "chalk";
 import { handle } from "./handle.js";
 
 export async function createPassword() {
-  console.log(chalk.bgGreen("password"))
+  console.log(chalk.green("password"))
 
   const password = await handle()
