+++ src/services/password/handle.js
@@ -1,24 +1,4 @@
-async function permittedCharacters() {
-  let permitted = []
-
-  if(process.env.UPPERCASE_LETTERS) {
-    permitted.push(..."ABCDEFGHIJKLMNOPQRSTUVWXYZ")
-  }
-
-  if(process.env.LOWCASE_LETTERS) {
-    permitted.push(..."abcdefghijklmnopqrstuvwxyz")
-  }
-
-  if(process.env.NUMBERS) {
-    permitted.push(..."0123456789")
-  }
-
-  if(process.env.SPECIAL_CHARACTERS) {
-    permitted.push(..."!@#$%&*+")
-  }
-
-  return permitted
-}
+import { permittedCharacters } from "./utils/permitted-characters.js"
 
 export async function handle() {
   let characters = []
@@ -28,10 +8,11 @@ export async function handle() {
   
   characters = await permittedCharacters()
 
-  for(let i = 0; i > passwordLength; i++) {
-    const index = Math.floor(Math.random() + characters.length)
-    password += characters[index]
+  for(let i = 0; i < passwordLength; i++) {
+    const index = Math.floor(Math.random() * characters.length)
 
-    return password
+    password += characters[index]  
   }
+
+  return password
 }
